extract medium filter and time periord filter and rename
