+++ src/Apps/Collect/Components/ArtworkGrid/index.tsx
@@ -1,14 +1,14 @@
-// import { LargeSelect, SmallSelect } from "@artsy/palette"
 import { ArtworkGrid_viewer } from "__generated__/ArtworkGrid_viewer.graphql"
 import { FilterState } from "Apps/Collect/FilterState"
 import { ContextConsumer } from "Artsy"
-import { SystemProps } from "Artsy/SystemContext"
 import React, { Component } from "react"
 import { createFragmentContainer, graphql } from "react-relay"
 import { Toggle } from "Styleguide/Components/Toggle"
 import { Subscribe } from "unstated"
 import { Responsive } from "Utils/Responsive"
+import { MediumFilter } from "../Filters/MediumFilter"
 import { PriceRange } from "../Filters/PriceRange"
+import { TimePeriodFilter } from "../Filters/TimePeriodFilter"
 import { CollectArtworkGridRefetchContainer as ArtworkFilter } from "./CollectArtworkFilterRefetch"
 
 import {
@@ -16,7 +16,6 @@ import {
   Checkbox,
   Flex,
   LargeSelect,
-  Radio,
   Sans,
   Separator,
   SmallSelect,
@@ -27,65 +26,6 @@ interface Props {
   hideTopBorder?: boolean
   viewer: ArtworkGrid_viewer
 }
-
-export const MediumRadios: React.SFC<{
-  filters: FilterState
-  mediums: ArtworkGrid_viewer["filter_artworks"]["aggregations"][0]["counts"]
-  mediator: SystemProps["mediator"]
-}> = ({ filters, mediums, mediator }) => (
-  <>
-    {mediums.map((medium, index) => {
-      const isSelected = filters.state.medium === medium.id
-
-      return (
-        <Radio
-          my={0.3}
-          selected={isSelected}
-          value={medium.id}
-          onSelect={({ selected }) => {
-            if (selected) {
-              return filters.setFilter("medium", medium.id, mediator)
-            } else {
-              return filters.unsetFilter("medium", mediator)
-            }
-          }}
-          key={index}
-        >
-          {medium.name}
-        </Radio>
-      )
-    })}
-  </>
-)
-
-export const TimePeriodRadios: React.SFC<{
-  filters: FilterState
-  mediator: SystemProps["mediator"]
-}> = ({ filters, mediator }) => (
-  <>
-    {allowedPeriods.map((timePeriod, index) => {
-      const isSelected = filters.state.major_periods[0] === timePeriod
-
-      return (
-        <Radio
-          my={0.3}
-          selected={isSelected}
-          value={timePeriod}
-          onSelect={({ selected }) => {
-            if (selected) {
-              return filters.setFilter("major_periods", timePeriod, mediator)
-            } else {
-              return filters.unsetFilter("major_periods", mediator)
-            }
-          }}
-          key={index}
-        >
-          {timePeriod}
-        </Radio>
-      )
-    })}
-  </>
-)
 
 class Filter extends Component<Props> {
   static defaultProps = {
@@ -195,7 +135,7 @@ class Filter extends Component<Props> {
                                 </Flex>
 
                                 <Toggle label="Medium" expanded>
-                                  <MediumRadios
+                                  <MediumFilter
                                     filters={filters}
                                     mediums={mediumAggregation.counts}
                                     mediator={mediator}
@@ -207,7 +147,7 @@ class Filter extends Component<Props> {
                                   }
                                   label="Time period"
                                 >
-                                  <TimePeriodRadios
+                                  <TimePeriodFilter
                                     filters={filters}
                                     mediator={mediator}
                                   />
@@ -331,21 +271,3 @@ export const ArtworkGridFragmentContainer = createFragmentContainer(
 )
 
 const Sidebar = Box
-
-const allowedPeriods = [
-  "2010",
-  "2000",
-  "1990",
-  "1980",
-  "1970",
-  "1960",
-  "1950",
-  "1940",
-  "1930",
-  "1920",
-  "1910",
-  "1900",
-  "Late 19th Century",
-  "Mid 19th Century",
-  "Early 19th Century",
-]
